Windows: Remove _once_cell_ dependency
