+++ Cargo.toml
@@ -11,7 +11,6 @@ readme="README.md"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3", features = ["minwindef", "mmsystem", "timeapi"] }
-once_cell = "1"
 
 [dev-dependencies]
 approx = "0.5"
